better circle, but still not good

+++ Engine/Graphics.h
@@ -57,7 +57,7 @@ public:
 		PutPixel( x,y,{ unsigned char( r ),unsigned char( g ),unsigned char( b ) } );
 	}
 	void PutPixel( int x,int y,Color c );
-	void DrawCircle(float xPos, float yPos, float r, Color c);
+	void DrawCircle(int xPos, int yPos, int r, Color c);
 	~Graphics();
 private:
 	Microsoft::WRL::ComPtr<IDXGISwapChain>				pSwapChain;
